Removal of more code replication.
